Dropping the "real cartridge" chapter
I want to write this, but... at the end of the day, I don't have the
knowledge for it, and I also think it is probably beyond the scope of this
guide anyway.

If someone else wants to write this and PR it in, I'll probably accept it,
but I don't plan to. If I ever make a cartridge release, my plans are to
work with Paul at IfiniteNesLives to either do the release, or buy some
cartridges to manage it myself.

Sorry if anyone was looking forward to this one!